az/requirements.txt
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b3
grpcio==1.14.2
grpcio-tools==1.14.2
pbr==4.2.0
protobuf==3.6.1
six==1.11.0
stevedore==1.29.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2

# comun
netifaces==0.10.7
# ml
matplotlib==3.0.2
mxnet==1.3.0.post0
# ml + math
numpy==1.14.6
# com
pymongo==3.7.2



